Changes to deprecated functions in Unreal Engine 5.4

+++ Source/glTFRuntime/Public/glTFRuntimeFunctionLibrary.h
@@ -8,7 +8,7 @@
 #include "glTFRuntimeFunctionLibrary.generated.h"
 
 DECLARE_DYNAMIC_DELEGATE_OneParam(FglTFRuntimeHttpResponse, UglTFRuntimeAsset*, Asset);
-DECLARE_DYNAMIC_DELEGATE_ThreeParams(FglTFRuntimeHttpProgress, const FglTFRuntimeConfig&, LoaderConfig, int32, BytesProcessed, int32, TotalBytes);
+DECLARE_DYNAMIC_DELEGATE_ThreeParams(FglTFRuntimeHttpProgress, const FglTFRuntimeConfig&, LoaderConfig, uint64, BytesProcessed, uint64, TotalBytes);
 
 /**
  *
